(editor) Extract visual sections and side-effects into PollenEditorBase
Move §6 Allergen icons, §7 Level circles and §8 Icon in ring into the
shared base as _renderAllergenIconsSection/_renderLevelCirclesSection/
_renderIconInRingSection, plus _inheritState() for the inherit/gap locals.
Extract the visual _updateConfig side effects (inherit-mode reset,
stroke-width to gap sync, allergen color-mode reset, icon_in_ring auto
thickness shift) into _applyVisualConfigSideEffects, and move _resetAll
to the base. The card editor now calls the shared methods; behavior is
unchanged. Lets the badge editor reuse the same visual config UI.
