+++ src/editor/base.js
@@ -9,18 +9,42 @@
 //   - _getAllergenDisplayName()
 //   - _hasSilamWeatherEntity()
 //   - _editorConfig(), _currentAllergens(), _currentNumLevels(), _thresholdParams()
+//   - _inheritState() — computed inherit/gap locals for §6/§7
 //   - _renderIntegrationSection() — §1 Integration & Location
 //   - _renderAllergensSection()   — §2 Allergens
 //   - _renderAppearanceSection()  — §5 Card appearance
+//   - _renderAllergenIconsSection() — §6 Allergen icons
+//   - _renderLevelCirclesSection()  — §7 Level circles
+//   - _renderIconInRingSection()    — §8 Icon in ring
+//   - _applyVisualConfigSideEffects(prop, value, config) — shared visual side-effects
+//   - _resetAll() — reset to stub defaults while preserving location keys
 //
 // Both editors extend this class. Section methods read from this._config
 // and call this._updateConfig / this._hass — both editors must expose those.
+//
+// _applyVisualConfigSideEffects CONTRACT:
+//   - Receives (prop, value, config) where config is a shallow copy of the
+//     current _config.
+//   - Returns { config, handled } where config is the (possibly mutated) config
+//     and handled=true means the caller should dispatch config-changed immediately
+//     and return (skipping the normal merge path).
+//   - The caller is responsible for writing config back to this._config,
+//     updating this._userConfig (if used), and dispatching config-changed.
+//   - The method does NOT touch this._config or this._userConfig directly.
+//     It operates purely on the passed-in config object and returns it.
+//   - The icon_in_ring branch needs the session flag this._thicknessAutoShifted;
+//     the caller must maintain that flag based on the returned state.
 
 import { LitElement, html } from "lit";
 import { t, detectLang } from "../i18n.js";
 import { normalize } from "../utils/normalize.js";
 import { slugify } from "../utils/slugify.js";
-import { LEVELS_DEFAULTS } from "../utils/levels-defaults.js";
+import {
+  LEVELS_DEFAULTS,
+  convertStrokeWidthToGap,
+  NORMAL_DEFAULT_THICKNESS,
+  ICON_IN_RING_DEFAULT_THICKNESS,
+} from "../utils/levels-defaults.js";
 import { getAllAdapterIds, getStubConfig } from "../adapter-registry.js";
 import {
   stubConfigDWD,
@@ -1237,6 +1261,24 @@ export class PollenEditorBase extends LitElement {
   }
 
   // ------------------------------------------------------------------
+  // Inherit-state helper (used by §6 and §7 section methods)
+  // ------------------------------------------------------------------
+
+  /**
+   * Computes the three locals that §6 and §7 templates depend on.
+   * Keeps the computation in one place so both sections stay in sync.
+   *
+   * @returns {{ inheritMode: string, gapSynced: boolean, gapDisabled: boolean }}
+   */
+  _inheritState() {
+    const c = this._editorConfig();
+    const inheritMode = c.levels_inherit_mode || "inherit_allergen";
+    const gapSynced = c.allergen_levels_gap_synced ?? true;
+    const gapDisabled = inheritMode === "inherit_allergen" && gapSynced;
+    return { inheritMode, gapSynced, gapDisabled };
+  }
+
+  // ------------------------------------------------------------------
   // §5 Card appearance section
   // ------------------------------------------------------------------
 
@@ -1314,5 +1356,977 @@ export class PollenEditorBase extends LitElement {
           </ha-formfield>
       </details>
     `;
+  }
+
+  // ------------------------------------------------------------------
+  // §6 Allergen icons section
+  // ------------------------------------------------------------------
+
+  _renderAllergenIconsSection() {
+    const c = this._editorConfig();
+    return html`
+      <!-- §6 Allergen icons -->
+      <details>
+        <summary>${this._t("summary_allergen_icons")}</summary>
+        <div class="section-helper">${this._t("helper_allergen_icons")}</div>
+        <ha-formfield
+          label="${this._t("allergen_color_mode") ||
+          "Allergen Color Mode"}"
+        >
+          <ha-selector
+            .hass=${this._hass}
+            .selector=${{
+              select: {
+                mode: "dropdown",
+                options: [
+                  {
+                    value: "default_colors",
+                    label:
+                      this._t("allergen_color_default_colors") ||
+                      "Default Colors",
+                  },
+                  {
+                    value: "custom",
+                    label:
+                      this._t("allergen_color_custom") || "Custom Colors",
+                  },
+                ],
+              },
+            }}
+            .value=${c.allergen_color_mode || "default_colors"}
+            @value-changed=${(e) => {
+              const v = e.detail?.value;
+              if (v !== undefined)
+                this._updateConfig("allergen_color_mode", v);
+            }}
+          ></ha-selector>
+        </ha-formfield>
+
+            ${c.allergen_color_mode === "custom"
+              ? html`
+                  <ha-formfield
+                    label="${this._t("allergen_colors") ||
+                    "Allergen Colors (by Level)"}"
+                  >
+                    <div
+                      style="display: flex; flex-direction: column; gap: 8px;"
+                    >
+                      ${(() => {
+                        const defaultAllergenColors =
+                          LEVELS_DEFAULTS.allergen_colors;
+                        const allergenColors =
+                          c.allergen_colors || defaultAllergenColors;
+
+                        return allergenColors.map(
+                          (col, i) => html`
+                            <div
+                              style="display: flex; align-items: center; gap: 8px;"
+                            >
+                              <span style="min-width: 60px;"
+                                >Level ${i}:</span
+                              >
+                              <input
+                                type="color"
+                                .value=${(() => {
+                                  if (i === 0 && col.includes("rgba")) {
+                                    return "#c8c8c8";
+                                  }
+                                  return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
+                                    col,
+                                  )
+                                    ? col
+                                    : "#000000";
+                                })()}
+                                @input=${(e) => {
+                                  const newColors = [...allergenColors];
+                                  newColors[i] = e.target.value;
+                                  this._updateConfig(
+                                    "allergen_colors",
+                                    newColors,
+                                  );
+                                }}
+                                style="width: 28px; height: 28px; border: none; background: none;"
+                              />
+                              <ha-textfield
+                                .value=${col}
+                                placeholder="${i === 0
+                                  ? this._t("allergen_empty_placeholder") ||
+                                    "rgba(200,200,200,0.15)"
+                                  : this._t("allergen_colors_placeholder") ||
+                                    "#ffcc00"}"
+                                @input=${(e) => {
+                                  const newColors = [...allergenColors];
+                                  newColors[i] = e.target.value;
+                                  this._updateConfig(
+                                    "allergen_colors",
+                                    newColors,
+                                  );
+                                }}
+                                style="width: 120px;"
+                              ></ha-textfield>
+                              <ha-button
+                                outlined
+                                title="${this._t("allergen_colors_reset") ||
+                                "Reset"}"
+                                @click=${() => {
+                                  const newColors = [...allergenColors];
+                                  newColors[i] =
+                                    LEVELS_DEFAULTS.allergen_colors[i];
+                                  this._updateConfig(
+                                    "allergen_colors",
+                                    newColors,
+                                  );
+                                }}
+                                style="margin-left: 8px;"
+                                >↺</ha-button
+                              >
+                            </div>
+                          `,
+                        );
+                      })()}
+                    </div>
+                  </ha-formfield>
+
+                  <ha-formfield
+                    label="${this._t("no_allergens_color") ||
+                    "No Allergens Color"}"
+                  >
+                    <div
+                      style="display: flex; align-items: center; gap: 8px;"
+                    >
+                      <input
+                        type="color"
+                        .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
+                          c.no_allergens_color ||
+                            LEVELS_DEFAULTS.no_allergens_color,
+                        )
+                          ? c.no_allergens_color ||
+                            LEVELS_DEFAULTS.no_allergens_color
+                          : "#a9cfe0"}
+                        @input=${(e) =>
+                          this._updateConfig(
+                            "no_allergens_color",
+                            e.target.value,
+                          )}
+                        style="width: 28px; height: 28px; border: none; background: none;"
+                      />
+                      <ha-textfield
+                        .value=${c.no_allergens_color ||
+                        LEVELS_DEFAULTS.no_allergens_color}
+                        placeholder="${this._t(
+                          "no_allergens_color_placeholder",
+                        ) || "#a9cfe0"}"
+                        @input=${(e) =>
+                          this._updateConfig(
+                            "no_allergens_color",
+                            e.target.value,
+                          )}
+                        style="width: 100px;"
+                      ></ha-textfield>
+                      <ha-button
+                        outlined
+                        title="${this._t("no_allergens_color_reset") ||
+                        "Reset"}"
+                        @click=${() =>
+                          this._updateConfig(
+                            "no_allergens_color",
+                            LEVELS_DEFAULTS.no_allergens_color,
+                          )}
+                        style="margin-left: 8px;"
+                        >↺</ha-button
+                      >
+                    </div>
+                  </ha-formfield>
+                `
+              : ""}
+        <ha-formfield
+          label="${this._t("allergen_outline_color")}"
+        >
+          <div style="display: flex; align-items: center; gap: 8px;">
+            <input
+              type="color"
+              .value=${(() => {
+                const color =
+                  c.allergen_outline_color ||
+                  LEVELS_DEFAULTS.levels_gap_color;
+                if (color.includes("rgba")) {
+                  return "#c8c8c8";
+                }
+                return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
+                  ? color
+                  : "#c8c8c8";
+              })()}
+              @input=${(e) =>
+                this._updateConfig(
+                  "allergen_outline_color",
+                  e.target.value,
+                )}
+              style="width: 28px; height: 28px; border: none; background: none;"
+            />
+            <ha-textfield
+              .value=${c.allergen_outline_color ||
+              LEVELS_DEFAULTS.levels_gap_color}
+              placeholder="${this._t(
+                "allergen_outline_placeholder",
+              ) || "rgba(200,200,200,1)"}"
+              @input=${(e) =>
+                this._updateConfig(
+                  "allergen_outline_color",
+                  e.target.value,
+                )}
+              style="width: 100px;"
+            ></ha-textfield>
+            <ha-button
+              outlined
+              title="${this._t("allergen_outline_reset") || "Reset"}"
+              @click=${() =>
+                this._updateConfig(
+                  "allergen_outline_color",
+                  LEVELS_DEFAULTS.levels_gap_color,
+                )}
+              style="margin-left: 8px;"
+              >↺</ha-button
+            >
+          </div>
+        </ha-formfield>
+        <ha-formfield
+          label="${this._t("allergen_stroke_color_synced")}"
+        >
+          <ha-checkbox
+            .checked=${c.allergen_stroke_color_synced ?? true}
+            @change=${(e) =>
+              this._updateConfig(
+                "allergen_stroke_color_synced",
+                e.target.checked,
+              )}
+          ></ha-checkbox>
+        </ha-formfield>
+        <ha-formfield
+          label="${this._t("allergen_stroke_width")}"
+        >
+          <ha-slider
+            min="0"
+            max="150"
+            step="5"
+            .value=${c.allergen_stroke_width ?? LEVELS_DEFAULTS.allergen_stroke_width}
+            @input=${(e) => {
+              const value = Number(e.target.value);
+              this._updateConfig("allergen_stroke_width", value);
+              const { inheritMode, gapSynced } = this._inheritState();
+              if (inheritMode === "inherit_allergen" && gapSynced) {
+                const levelGap = convertStrokeWidthToGap(value);
+                this._updateConfig("levels_gap", levelGap);
+              }
+            }}
+            style="width: 120px;"
+          ></ha-slider>
+          <ha-textfield
+            type="number"
+            min="0"
+            max="150"
+            step="5"
+            .value=${c.allergen_stroke_width ?? LEVELS_DEFAULTS.allergen_stroke_width}
+            @input=${(e) => {
+              const value = e.target.value === "" ? LEVELS_DEFAULTS.allergen_stroke_width : Number(e.target.value);
+              this._updateConfig("allergen_stroke_width", value);
+              const { inheritMode, gapSynced } = this._inheritState();
+              if (inheritMode === "inherit_allergen" && gapSynced) {
+                const levelGap = convertStrokeWidthToGap(value);
+                this._updateConfig("levels_gap", levelGap);
+              }
+            }}
+            style="width: 80px;"
+          ></ha-textfield>
+          <ha-button
+            outlined
+            title="${this._t("allergen_stroke_width_reset") || "Reset"}"
+            @click=${() =>
+              this._updateConfig(
+                "allergen_stroke_width",
+                LEVELS_DEFAULTS.allergen_stroke_width,
+              )}
+            style="margin-left: 8px;"
+            >↺</ha-button
+          >
+        </ha-formfield>
+        <div class="field-helper">${this._t("helper_allergen_stroke_width")}</div>
+      </details>
+    `;
+  }
+
+  // ------------------------------------------------------------------
+  // §7 Level circles section
+  // ------------------------------------------------------------------
+
+  _renderLevelCirclesSection() {
+    const c = this._editorConfig();
+    const { inheritMode, gapSynced, gapDisabled } = this._inheritState();
+    return html`
+      <!-- §7 Level circles -->
+      <details>
+        <summary>${this._t("summary_level_circles")}</summary>
+        <div class="section-helper">${this._t("helper_level_circles")}</div>
+        <ha-formfield
+          label="${this._t("levels_inherit_mode")}"
+        >
+          <ha-selector
+            .hass=${this._hass}
+            .selector=${{
+              select: {
+                mode: "dropdown",
+                options: [
+                  {
+                    value: "inherit_allergen",
+                    label:
+                      this._t("levels_inherit_allergen") ||
+                      "Inherit from Allergen Colors",
+                  },
+                  {
+                    value: "custom",
+                    label:
+                      this._t("levels_custom") || "Use Custom Level Colors",
+                  },
+                ],
+              },
+            }}
+            .value=${c.levels_inherit_mode || "inherit_allergen"}
+            @value-changed=${(e) => {
+              const v = e.detail?.value;
+              if (v !== undefined) this._updateConfig("levels_inherit_mode", v);
+            }}
+          ></ha-selector>
+        </ha-formfield>
+
+        ${inheritMode === "inherit_allergen"
+          ? html`
+              <ha-formfield
+                label="${this._t("allergen_levels_gap_synced")}"
+              >
+                <ha-checkbox
+                  .checked=${c.allergen_levels_gap_synced ?? true}
+                  @change=${(e) =>
+                    this._updateConfig(
+                      "allergen_levels_gap_synced",
+                      e.target.checked,
+                    )}
+                ></ha-checkbox>
+              </ha-formfield>
+              <div class="field-helper">${this._t("helper_allergen_levels_gap_synced")}</div>
+            `
+          : ""}
+
+        ${inheritMode === "custom"
+          ? html`
+              <ha-formfield label="${this._t("levels_colors")}">
+                <div style="display: flex; flex-direction: column; gap: 8px;">
+                  ${c.levels_colors.map(
+                    (col, i) => html`
+                      <div style="display: flex; align-items: center; gap: 8px;">
+                        <input
+                          type="color"
+                          .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(col)
+                            ? col
+                            : "#000000"}
+                          @input=${(e) => {
+                            const newColors = [...c.levels_colors];
+                            newColors[i] = e.target.value;
+                            this._updateConfig("levels_colors", newColors);
+                          }}
+                          style="width: 28px; height: 28px; border: none; background: none;"
+                        />
+                        <ha-textfield
+                          .value=${col}
+                          placeholder="${this._t("levels_colors_placeholder")}"
+                          @input=${(e) => {
+                            const newColors = [...c.levels_colors];
+                            newColors[i] = e.target.value;
+                            this._updateConfig("levels_colors", newColors);
+                          }}
+                          style="width: 100px;"
+                        ></ha-textfield>
+                        <ha-button
+                          outlined
+                          title="${this._t("levels_reset")}"
+                          @click=${() => {
+                            const newColors = [...c.levels_colors];
+                            newColors[i] = LEVELS_DEFAULTS.levels_colors[i];
+                            this._updateConfig("levels_colors", newColors);
+                          }}
+                          style="margin-left: 8px;"
+                          >↺</ha-button
+                        >
+                      </div>
+                    `,
+                  )}
+                </div>
+              </ha-formfield>
+
+              <ha-formfield label="${this._t("levels_empty_color")}">
+                <div style="display: flex; align-items: center; gap: 8px;">
+                  <input
+                    type="color"
+                    .value=${(() => {
+                      const color =
+                        c.levels_empty_color ||
+                        LEVELS_DEFAULTS.levels_empty_color;
+                      if (color.includes("rgba")) {
+                        return "#c8c8c8";
+                      }
+                      return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
+                        ? color
+                        : "#c8c8c8";
+                    })()}
+                    @input=${(e) =>
+                      this._updateConfig("levels_empty_color", e.target.value)}
+                    style="width: 28px; height: 28px; border: none; background: none;"
+                  />
+                  <ha-textfield
+                    .value=${c.levels_empty_color}
+                    placeholder="${this._t("levels_colors_placeholder")}"
+                    @input=${(e) =>
+                      this._updateConfig("levels_empty_color", e.target.value)}
+                    style="width: 100px;"
+                  ></ha-textfield>
+                  <ha-button
+                    outlined
+                    title="${this._t("levels_reset")}"
+                    @click=${() =>
+                      this._updateConfig(
+                        "levels_empty_color",
+                        LEVELS_DEFAULTS.levels_empty_color,
+                      )}
+                    style="margin-left: 8px;"
+                    >↺</ha-button
+                  >
+                </div>
+              </ha-formfield>
+            `
+          : ""}
+
+        <ha-formfield label="${this._t("levels_thickness")}">
+          <ha-slider
+            min="10"
+            max="90"
+            step="1"
+            .value=${c.levels_thickness}
+            @input=${(e) =>
+              this._updateConfig("levels_thickness", Number(e.target.value))}
+            style="width: 120px;"
+          ></ha-slider>
+          <ha-textfield
+            type="number"
+            .value=${c.levels_thickness}
+            @input=${(e) =>
+              this._updateConfig("levels_thickness", Number(e.target.value))}
+            style="width: 80px;"
+          ></ha-textfield>
+          <ha-button
+            outlined
+            title="${this._t("levels_reset")}"
+            @click=${() =>
+              this._updateConfig("levels_thickness", LEVELS_DEFAULTS.levels_thickness)}
+            style="margin-left: 8px;"
+            >↺</ha-button
+          >
+        </ha-formfield>
+
+        <ha-formfield
+          label="${this._t("levels_gap")}"
+          .disabled=${gapDisabled}
+        >
+          <ha-slider
+            min="0"
+            max="20"
+            step="1"
+            .value=${c.levels_gap}
+            .disabled=${gapDisabled}
+            @input=${(e) =>
+              this._updateConfig("levels_gap", Number(e.target.value))}
+            style="width: 120px;"
+          ></ha-slider>
+          <ha-textfield
+            type="number"
+            .value=${c.levels_gap}
+            .disabled=${gapDisabled}
+            @input=${(e) =>
+              this._updateConfig("levels_gap", Number(e.target.value))}
+            style="width: 80px;"
+          ></ha-textfield>
+          <ha-button
+            outlined
+            title="${this._t("levels_reset")}"
+            .disabled=${gapDisabled}
+            @click=${() =>
+              this._updateConfig("levels_gap", LEVELS_DEFAULTS.levels_gap)}
+            style="margin-left: 8px;"
+            >↺</ha-button
+          >
+        </ha-formfield>
+        <div class="field-helper">
+          ${gapDisabled
+            ? this._t("helper_levels_gap_synced")
+            : this._t("helper_levels_gap_unsynced")}
+        </div>
+
+        ${inheritMode === "custom" || !gapSynced
+          ? html`
+              <ha-formfield label="${this._t("levels_gap_color")}">
+                <div style="display: flex; align-items: center; gap: 8px;">
+                  <input
+                    type="color"
+                    .value=${(() => {
+                      const color =
+                        c.levels_gap_color || LEVELS_DEFAULTS.levels_gap_color;
+                      if (color.includes("rgba")) {
+                        return "#c8c8c8";
+                      }
+                      return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
+                        ? color
+                        : "#c8c8c8";
+                    })()}
+                    @input=${(e) =>
+                      this._updateConfig("levels_gap_color", e.target.value)}
+                    style="width: 28px; height: 28px; border: none; background: none;"
+                  />
+                  <ha-textfield
+                    .value=${c.levels_gap_color}
+                    placeholder="${this._t("levels_colors_placeholder")}"
+                    @input=${(e) =>
+                      this._updateConfig("levels_gap_color", e.target.value)}
+                    style="width: 100px;"
+                  ></ha-textfield>
+                  <ha-button
+                    outlined
+                    title="${this._t("levels_reset")}"
+                    @click=${() =>
+                      this._updateConfig(
+                        "levels_gap_color",
+                        LEVELS_DEFAULTS.levels_gap_color,
+                      )}
+                    style="margin-left: 8px;"
+                    >↺</ha-button
+                  >
+                </div>
+              </ha-formfield>
+            `
+          : ""}
+
+        <ha-formfield label="${this._t("show_value_numeric_in_circle")}">
+          <ha-switch
+            .checked=${c.show_value_numeric_in_circle}
+            @change=${(e) =>
+              this._updateConfig(
+                "show_value_numeric_in_circle",
+                e.target.checked,
+              )}
+          ></ha-switch>
+        </ha-formfield>
+        <div class="field-helper">${this._t("helper_show_value_numeric_in_circle")}</div>
+
+        <ha-formfield label="${this._t("levels_text_weight")}">
+          <ha-selector
+            .hass=${this._hass}
+            .selector=${{
+              select: {
+                mode: "dropdown",
+                options: [
+                  { value: "normal", label: "Normal" },
+                  { value: "500", label: "Medium" },
+                  { value: "bold", label: "Bold" },
+                ],
+              },
+            }}
+            .value=${c.levels_text_weight || "normal"}
+            @value-changed=${(e) => {
+              const v = e.detail?.value;
+              if (v !== undefined) this._updateConfig("levels_text_weight", v);
+            }}
+          ></ha-selector>
+        </ha-formfield>
+
+        <ha-formfield label="${this._t("levels_text_size")}">
+          <ha-slider
+            min="0.1"
+            max="0.5"
+            step="0.05"
+            .value=${c.levels_text_size || 0.3}
+            @input=${(e) =>
+              this._updateConfig("levels_text_size", Number(e.target.value))}
+            style="width: 120px;"
+          ></ha-slider>
+          <ha-textfield
+            type="number"
+            .value=${c.levels_text_size || 0.3}
+            @input=${(e) =>
+              this._updateConfig("levels_text_size", Number(e.target.value))}
+            style="width: 80px;"
+          ></ha-textfield>
+        </ha-formfield>
+
+        <ha-formfield label="${this._t("levels_icon_ratio")}">
+          <ha-slider
+            min="0.1"
+            max="2"
+            step="0.05"
+            .value=${c.levels_icon_ratio || 1}
+            @input=${(e) =>
+              this._updateConfig("levels_icon_ratio", Number(e.target.value))}
+            style="width: 120px;"
+          ></ha-slider>
+          <ha-textfield
+            type="number"
+            .value=${c.levels_icon_ratio || 1}
+            @input=${(e) =>
+              this._updateConfig("levels_icon_ratio", Number(e.target.value))}
+            style="width: 80px;"
+          ></ha-textfield>
+        </ha-formfield>
+
+        <ha-formfield label="${this._t("levels_text_color")}">
+          <div style="display: flex; align-items: center; gap: 8px;">
+            <input
+              type="color"
+              .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
+                c.levels_text_color || "",
+              )
+                ? c.levels_text_color
+                : "#000000"}
+              @input=${(e) =>
+                this._updateConfig("levels_text_color", e.target.value)}
+              style="width: 28px; height: 28px; border: none; background: none;"
+            />
+            <ha-textfield
+              .value=${c.levels_text_color || ""}
+              placeholder="var(--primary-text-color)"
+              @input=${(e) =>
+                this._updateConfig("levels_text_color", e.target.value)}
+              style="width: 100px;"
+            ></ha-textfield>
+          </div>
+        </ha-formfield>
+      </details>
+    `;
+  }
+
+  // ------------------------------------------------------------------
+  // §8 Icon in ring section
+  // ------------------------------------------------------------------
+
+  _renderIconInRingSection() {
+    const c = this._editorConfig();
+    return html`
+      <!-- §8 Icon in ring -->
+      <details>
+        <summary>${this._t("summary_icon_in_ring")}</summary>
+        <div class="section-helper">${this._t("helper_icon_in_ring")}</div>
+        <ha-formfield
+          label="${this._t("icon_in_ring")}"
+        >
+          <ha-checkbox
+            .checked=${c.icon_in_ring === true}
+            @change=${(e) =>
+              this._updateConfig("icon_in_ring", e.target.checked)}
+          ></ha-checkbox>
+        </ha-formfield>
+        <ha-formfield
+          label="${this._t("icon_in_ring_size_ratio")}"
+        >
+          <div style="display: flex; align-items: center; gap: 8px;">
+            <ha-slider
+              min="0.2"
+              max="0.9"
+              step="0.05"
+              .value=${c.icon_in_ring_size_ratio ??
+              LEVELS_DEFAULTS.icon_in_ring_size_ratio}
+              @change=${(e) =>
+                this._updateConfig(
+                  "icon_in_ring_size_ratio",
+                  Number(e.target.value),
+                )}
+            ></ha-slider>
+            <ha-textfield
+              type="number"
+              min="0.2"
+              max="0.9"
+              step="0.05"
+              .value=${c.icon_in_ring_size_ratio ??
+              LEVELS_DEFAULTS.icon_in_ring_size_ratio}
+              @change=${(e) =>
+                this._updateConfig(
+                  "icon_in_ring_size_ratio",
+                  Number(e.target.value),
+                )}
+              style="width: 80px;"
+            ></ha-textfield>
+          </div>
+        </ha-formfield>
+        <ha-formfield
+          label="${this._t("icon_in_ring_color_mode")}"
+        >
+          <ha-selector
+            .hass=${this._hass}
+            .selector=${{
+              select: {
+                mode: "dropdown",
+                options: [
+                  {
+                    value: "static",
+                    label:
+                      this._t("icon_in_ring_color_static") ||
+                      "Static color",
+                  },
+                  {
+                    value: "follow_level",
+                    label:
+                      this._t("icon_in_ring_color_follow") ||
+                      "Follow level color",
+                  },
+                ],
+              },
+            }}
+            .value=${c.icon_in_ring_color_mode || "static"}
+            @value-changed=${(e) => {
+              const v = e.detail?.value;
+              if (v !== undefined)
+                this._updateConfig("icon_in_ring_color_mode", v);
+            }}
+          ></ha-selector>
+        </ha-formfield>
+        ${(c.icon_in_ring_color_mode || "static") === "static"
+          ? (() => {
+              const currentColor =
+                typeof c.icon_in_ring_static_color === "string"
+                  ? c.icon_in_ring_static_color
+                  : LEVELS_DEFAULTS.icon_in_ring_static_color;
+              const isHex = /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
+                currentColor,
+              );
+              return html`
+                <ha-formfield
+                  label="${this._t("icon_in_ring_static_color")}"
+                >
+                  <div
+                    style="display: flex; align-items: center; gap: 8px;"
+                  >
+                    ${isHex
+                      ? html`<input
+                          type="color"
+                          .value=${currentColor}
+                          @input=${(e) =>
+                            this._updateConfig(
+                              "icon_in_ring_static_color",
+                              e.target.value,
+                            )}
+                          style="width: 28px; height: 28px; border: none; background: none;"
+                        />`
+                      : html`<div
+                          title="${this._t(
+                            "icon_in_ring_static_color_var_hint",
+                          ) ||
+                          "Non-hex value (e.g. CSS variable); use the text field to edit"}"
+                          style="width: 28px; height: 28px; border: 1px dashed var(--divider-color); border-radius: 4px; display: flex; align-items: center; justify-content: center; font-size: 10px; color: var(--secondary-text-color);"
+                        >
+                          var
+                        </div>`}
+                    <ha-textfield
+                      .value=${currentColor}
+                      placeholder="${LEVELS_DEFAULTS.icon_in_ring_static_color}"
+                      @input=${(e) =>
+                        this._updateConfig(
+                          "icon_in_ring_static_color",
+                          e.target.value,
+                        )}
+                      style="width: 220px;"
+                    ></ha-textfield>
+                  </div>
+                </ha-formfield>
+              `;
+            })()
+          : ""}
+      </details>
+    `;
+  }
+
+  // ------------------------------------------------------------------
+  // Shared visual config side-effects
+  // ------------------------------------------------------------------
+
+  /**
+   * Applies visual config side-effects shared between the card editor and
+   * the badge editor. Operates on a passed-in config object (shallow copy
+   * of _config) and returns the (possibly mutated) config plus a `handled`
+   * flag.
+   *
+   * When handled=true, the caller MUST dispatch config-changed immediately
+   * and return (skip normal merge). The caller owns _config, _userConfig,
+   * and this._thicknessAutoShifted — this method does not touch them.
+   *
+   * @param {string} prop
+   * @param {*} value
+   * @param {object} config  Shallow copy of current _config to mutate
+   * @returns {{ config: object, handled: boolean, thicknessAutoShifted: boolean|null }}
+   *   thicknessAutoShifted is non-null only when the icon_in_ring branch fires;
+   *   null means leave this._thicknessAutoShifted unchanged.
+   */
+  _applyVisualConfigSideEffects(prop, value, config) {
+    // icon_in_ring auto-toggle for levels_thickness.
+    if (prop === "icon_in_ring") {
+      const prev = config.icon_in_ring === true;
+      const next = value === true;
+      if (prev !== next) {
+        const currentThickness =
+          config.levels_thickness ?? LEVELS_DEFAULTS.levels_thickness;
+        const newConfig = { ...config, icon_in_ring: next };
+        let thicknessAutoShifted = this._thicknessAutoShifted || false;
+        if (next && currentThickness === NORMAL_DEFAULT_THICKNESS) {
+          newConfig.levels_thickness = ICON_IN_RING_DEFAULT_THICKNESS;
+          thicknessAutoShifted = true;
+        } else if (
+          !next &&
+          thicknessAutoShifted &&
+          currentThickness === ICON_IN_RING_DEFAULT_THICKNESS
+        ) {
+          newConfig.levels_thickness = NORMAL_DEFAULT_THICKNESS;
+          thicknessAutoShifted = false;
+        }
+        return { config: newConfig, handled: true, thicknessAutoShifted };
+      }
+    }
+
+    // levels_inherit_mode: reset/sync related color properties.
+    if (prop === "levels_inherit_mode") {
+      if (value === "custom" && config.levels_inherit_mode !== "custom") {
+        const newConfig = {
+          ...config,
+          levels_inherit_mode: value,
+          levels_gap: LEVELS_DEFAULTS.levels_gap,
+          levels_colors: LEVELS_DEFAULTS.levels_colors,
+          levels_empty_color: LEVELS_DEFAULTS.levels_empty_color,
+          levels_gap_color: LEVELS_DEFAULTS.levels_gap_color,
+        };
+        return { config: newConfig, handled: true, thicknessAutoShifted: null };
+      } else if (
+        value === "inherit_allergen" &&
+        config.levels_inherit_mode === "custom"
+      ) {
+        const currentStrokeWidth =
+          config.allergen_stroke_width || LEVELS_DEFAULTS.allergen_stroke_width;
+        const syncedGap = convertStrokeWidthToGap(currentStrokeWidth);
+        const currentAllergenColors =
+          config.allergen_colors || LEVELS_DEFAULTS.allergen_colors;
+        const syncedEmptyColor =
+          currentAllergenColors[0] || LEVELS_DEFAULTS.levels_empty_color;
+        const newConfig = {
+          ...config,
+          levels_inherit_mode: value,
+          levels_gap: syncedGap,
+          levels_empty_color: syncedEmptyColor,
+          allergen_levels_gap_synced: true,
+        };
+        return { config: newConfig, handled: true, thicknessAutoShifted: null };
+      }
+    }
+
+    // allergen_colors: sync levels_empty_color when in inherit mode.
+    if (prop === "allergen_colors" && Array.isArray(value)) {
+      const newConfig = { ...config, allergen_colors: value };
+      if (
+        (config.levels_inherit_mode || "inherit_allergen") === "inherit_allergen"
+      ) {
+        if (value[0]) {
+          newConfig.levels_empty_color = value[0];
+        }
+      }
+      return { config: newConfig, handled: true, thicknessAutoShifted: null };
+    }
+
+    // allergen_stroke_width reset: sync levels_gap.
+    if (
+      prop === "allergen_stroke_width" &&
+      value === LEVELS_DEFAULTS.allergen_stroke_width
+    ) {
+      const newConfig = { ...config, allergen_stroke_width: value };
+      if (
+        (config.levels_inherit_mode || "inherit_allergen") ===
+          "inherit_allergen" &&
+        (config.allergen_levels_gap_synced ?? true)
+      ) {
+        const levelGap = convertStrokeWidthToGap(value);
+        newConfig.levels_gap = levelGap;
+      }
+      return { config: newConfig, handled: true, thicknessAutoShifted: null };
+    }
+
+    // Level visual property resets: ensure chart re-rendering.
+    if (
+      (prop === "levels_thickness" ||
+        prop === "levels_gap" ||
+        prop === "levels_colors" ||
+        prop === "levels_empty_color" ||
+        prop === "levels_gap_color") &&
+      value === LEVELS_DEFAULTS[prop]
+    ) {
+      const newConfig = { ...config, [prop]: value };
+      return { config: newConfig, handled: true, thicknessAutoShifted: null };
+    }
+
+    // allergen_color_mode reset to default_colors.
+    if (
+      prop === "allergen_color_mode" &&
+      value === "default_colors" &&
+      config.allergen_color_mode === "custom"
+    ) {
+      const newConfig = {
+        ...config,
+        allergen_color_mode: value,
+        allergen_colors: LEVELS_DEFAULTS.allergen_colors,
+        allergen_outline_color: LEVELS_DEFAULTS.levels_gap_color,
+        no_allergens_color: LEVELS_DEFAULTS.no_allergens_color,
+      };
+      return { config: newConfig, handled: true, thicknessAutoShifted: null };
+    }
+
+    // levels_thickness manual edit after auto-shift: signal caller to clear the flag.
+    if (prop === "levels_thickness" && this._thicknessAutoShifted) {
+      return { config, handled: false, thicknessAutoShifted: false };
+    }
+
+    return { config, handled: false, thicknessAutoShifted: null };
+  }
+
+  // ------------------------------------------------------------------
+  // Reset to stub defaults
+  // ------------------------------------------------------------------
+
+  /**
+   * Resets the card/badge config to stub defaults while preserving location
+   * identity keys (city, location, region_id, entity_prefix, entity_suffix,
+   * entity_weather, type). Both editors inherit this; the badge editor works
+   * correctly because badge_* keys are absent from the stub and will revert
+   * to undefined (the badge element re-applies its own defaults on setConfig).
+   */
+  _resetAll() {
+    const KEEP = [
+      "city", "location", "region_id",
+      "entity_prefix", "entity_suffix", "entity_weather",
+      "type",
+    ];
+    const integration = this._config?.integration ?? "pp";
+    const stub = getStubConfig(integration) || {};
+    const preserved = {};
+    for (const k of KEEP) {
+      if (this._config?.[k] !== undefined) preserved[k] = this._config[k];
+    }
+    const fresh = { ...stub, ...preserved, integration };
+    // Let setConfig (defined by the subclass) normalise and commit the reset.
+    this.setConfig(fresh);
+    // Notify HA that config changed.
+    this.dispatchEvent(
+      new CustomEvent("config-changed", {
+        detail: { config: this._config },
+        bubbles: true,
+        composed: true,
+      }),
+    );
   }
 }
