+++ src/pollenprognos-editor.js
@@ -4,8 +4,6 @@ import { t, detectLang, SUPPORTED_LOCALES } from "./i18n.js";
 import { deepEqual } from "./utils/confcompare.js";
 import {
   LEVELS_DEFAULTS,
-  convertStrokeWidthToGap,
-  NORMAL_DEFAULT_THICKNESS,
   ICON_IN_RING_DEFAULT_THICKNESS,
 } from "./utils/levels-defaults.js";
 import { COSMETIC_FIELDS } from "./constants.js";
@@ -40,12 +38,12 @@ import {
 import silamAllergenMap from "./adapters/silam_allergen_map.json" assert { type: "json" };
 
 class PollenPrognosCardEditor extends PollenEditorBase {
+  // _resetAll is inherited from PollenEditorBase. The card editor overrides
+  // to also clear _userConfig before delegating to the base implementation.
   _resetAll() {
     if (this.debug) console.debug("[Editor] resetAll");
     this._userConfig = {};
-    // Sätt integration och type explicit
-    const integration = this._config?.integration ?? "pp";
-    this.setConfig({ integration, type: "custom:pollenprognos-card" });
+    super._resetAll();
   }
 
   _resetPhrases(lang) {
@@ -1434,233 +1432,49 @@ class PollenPrognosCardEditor extends PollenEditorBase {
       return;
     }
 
-    // Icon-in-ring auto-toggle for levels_thickness (#227).
-    // When the user enables icon_in_ring with the normal-mode default
-    // thickness still in place, snap to the icon-in-ring default so the
-    // icon has room. Reverse on disable — but only if this editor
-    // session actually performed the enable-time auto-shift. This
-    // protects a user who manually set levels_thickness to 35 (or 60)
-    // before the toggle: the swap won't overwrite their value just
-    // because it happens to equal the other mode's default.
-    if (prop === "icon_in_ring") {
-      const prev = this._config?.icon_in_ring === true;
-      const next = value === true;
-      if (prev !== next) {
-        const currentThickness =
-          this._config?.levels_thickness ??
-          LEVELS_DEFAULTS.levels_thickness;
-        const newConfig = { ...this._config, icon_in_ring: next };
-        if (next && currentThickness === NORMAL_DEFAULT_THICKNESS) {
-          newConfig.levels_thickness = ICON_IN_RING_DEFAULT_THICKNESS;
-          this._userConfig.levels_thickness = ICON_IN_RING_DEFAULT_THICKNESS;
-          this._thicknessAutoShifted = true;
-        } else if (
-          !next &&
-          this._thicknessAutoShifted &&
-          currentThickness === ICON_IN_RING_DEFAULT_THICKNESS
-        ) {
-          newConfig.levels_thickness = NORMAL_DEFAULT_THICKNESS;
-          this._userConfig.levels_thickness = NORMAL_DEFAULT_THICKNESS;
-          this._thicknessAutoShifted = false;
+    // Delegate visual side-effects to the shared base helper. The helper
+    // returns handled=true when it has built the final config and the caller
+    // should dispatch immediately; it also returns thicknessAutoShifted so
+    // the card editor can maintain that session flag.
+    {
+      const result = this._applyVisualConfigSideEffects(
+        prop, value, { ...this._config }
+      );
+      if (result.thicknessAutoShifted !== null) {
+        this._thicknessAutoShifted = result.thicknessAutoShifted;
+      }
+      if (result.handled) {
+        // Sync _userConfig for all mutated keys so set hass() round-trips
+        // don't lose the changes.
+        const prev = this._config;
+        this._config = result.config;
+        for (const k of Object.keys(result.config)) {
+          if (result.config[k] !== prev[k]) {
+            this._userConfig[k] = result.config[k];
+          }
         }
-        this._config = newConfig;
-        this._userConfig.icon_in_ring = next;
-        this.dispatchEvent(
-          new CustomEvent("config-changed", {
-            detail: { config: newConfig },
-            bubbles: true,
-            composed: true,
-          }),
-        );
-        return;
-      }
-    }
-
-    // If the user manually edits levels_thickness after an auto-shift,
-    // clear the flag so a future icon_in_ring toggle doesn't try to
-    // restore a value the user has since overridden.
-    if (prop === "levels_thickness" && this._thicknessAutoShifted) {
-      this._thicknessAutoShifted = false;
-    }
-
-    // Handle levels_inherit_mode changes - reset gap and sync when needed
-    if (prop === "levels_inherit_mode") {
-      if (value === "custom" && this._config.levels_inherit_mode !== "custom") {
-        // Switching to custom - reset gap to default
-        const newConfig = {
-          ...this._config,
-          levels_inherit_mode: value,
-          levels_gap: LEVELS_DEFAULTS.levels_gap,
-          levels_colors: LEVELS_DEFAULTS.levels_colors,
-          levels_empty_color: LEVELS_DEFAULTS.levels_empty_color,
-          levels_gap_color: LEVELS_DEFAULTS.levels_gap_color,
-        };
-        this._config = newConfig;
-        this._userConfig.levels_inherit_mode = value;
-        delete this._userConfig.levels_gap;
-        delete this._userConfig.levels_colors;
-        delete this._userConfig.levels_empty_color;
-        delete this._userConfig.levels_gap_color;
-
-        this.dispatchEvent(
-          new CustomEvent("config-changed", {
-            detail: { config: newConfig },
-            bubbles: true,
-            composed: true,
-          }),
-        );
-        return;
-      } else if (
-        value === "inherit_allergen" &&
-        this._config.levels_inherit_mode === "custom"
-      ) {
-        // Switching back to inherit - sync gap with current stroke width and empty color
-        const currentStrokeWidth =
-          this._config.allergen_stroke_width ||
-          LEVELS_DEFAULTS.allergen_stroke_width;
-        const syncedGap = convertStrokeWidthToGap(currentStrokeWidth);
-
-        // Also sync empty color from allergen colors[0]
-        const currentAllergenColors =
-          this._config.allergen_colors || LEVELS_DEFAULTS.allergen_colors;
-        const syncedEmptyColor =
-          currentAllergenColors[0] || LEVELS_DEFAULTS.levels_empty_color;
-
-        const newConfig = {
-          ...this._config,
-          levels_inherit_mode: value,
-          levels_gap: syncedGap,
-          levels_empty_color: syncedEmptyColor,
-          allergen_levels_gap_synced: true, // Enable sync when switching to inherit mode
-        };
-        this._config = newConfig;
-        this._userConfig.levels_inherit_mode = value;
-        this._userConfig.levels_gap = syncedGap;
-        this._userConfig.levels_empty_color = syncedEmptyColor;
-        this._userConfig.allergen_levels_gap_synced = true;
-
-        this.dispatchEvent(
-          new CustomEvent("config-changed", {
-            detail: { config: newConfig },
-            bubbles: true,
-            composed: true,
-          }),
-        );
-        return;
-      }
-    }
-
-    // Handle allergen colors changes - sync with levels empty color if inheriting
-    if (prop === "allergen_colors" && Array.isArray(value)) {
-      const newConfig = { ...this._config, allergen_colors: value };
-      this._userConfig.allergen_colors = value;
-
-      // Sync level 0 (empty) color with levels_empty_color if levels inherit from allergen
-      if (
-        (this._config.levels_inherit_mode || "inherit_allergen") ===
-        "inherit_allergen"
-      ) {
-        // allergen_colors[0] is the empty/level 0 color
-        if (value[0]) {
-          newConfig.levels_empty_color = value[0];
-          this._userConfig.levels_empty_color = value[0];
+        // allergen_color_mode reset clears related userConfig keys
+        if (prop === "allergen_color_mode" && value === "default_colors") {
+          delete this._userConfig.allergen_colors;
+          delete this._userConfig.allergen_outline_color;
+          delete this._userConfig.no_allergens_color;
         }
+        // levels_inherit_mode→custom clears color overrides from userConfig
+        if (prop === "levels_inherit_mode" && value === "custom") {
+          delete this._userConfig.levels_gap;
+          delete this._userConfig.levels_colors;
+          delete this._userConfig.levels_empty_color;
+          delete this._userConfig.levels_gap_color;
+        }
+        this.dispatchEvent(
+          new CustomEvent("config-changed", {
+            detail: { config: result.config },
+            bubbles: true,
+            composed: true,
+          }),
+        );
+        return;
       }
-
-      this._config = newConfig;
-
-      this.dispatchEvent(
-        new CustomEvent("config-changed", {
-          detail: { config: newConfig },
-          bubbles: true,
-          composed: true,
-        }),
-      );
-      return;
-    }
-
-    // Handle allergen stroke width reset - sync with levels gap if inheriting and synced
-    if (
-      prop === "allergen_stroke_width" &&
-      value === LEVELS_DEFAULTS.allergen_stroke_width
-    ) {
-      const newConfig = { ...this._config, allergen_stroke_width: value };
-      this._userConfig.allergen_stroke_width = value;
-
-      // Sync with level circle gap only if levels inherit from allergen AND gap sync is enabled
-      if (
-        (this._config.levels_inherit_mode || "inherit_allergen") ===
-          "inherit_allergen" &&
-        (this._config.allergen_levels_gap_synced ?? true)
-      ) {
-        const levelGap = convertStrokeWidthToGap(value);
-        newConfig.levels_gap = levelGap;
-        this._userConfig.levels_gap = levelGap;
-      }
-
-      this._config = newConfig;
-
-      this.dispatchEvent(
-        new CustomEvent("config-changed", {
-          detail: { config: newConfig },
-          bubbles: true,
-          composed: true,
-        }),
-      );
-      return;
-    }
-
-    // Handle level settings resets - ensure chart sync for visual properties
-    if (
-      (prop === "levels_thickness" ||
-        prop === "levels_gap" ||
-        prop === "levels_colors" ||
-        prop === "levels_empty_color" ||
-        prop === "levels_gap_color") &&
-      value === LEVELS_DEFAULTS[prop]
-    ) {
-      // For level visual properties, create a config change event to ensure chart re-rendering
-      const newConfig = { ...this._config, [prop]: value };
-      this._config = newConfig;
-      this._userConfig[prop] = value;
-
-      this.dispatchEvent(
-        new CustomEvent("config-changed", {
-          detail: { config: newConfig },
-          bubbles: true,
-          composed: true,
-        }),
-      );
-      return;
-    }
-
-    // Handle allergen color mode changes - reset colors when switching to default
-    if (
-      prop === "allergen_color_mode" &&
-      value === "default_colors" &&
-      this._config.allergen_color_mode === "custom"
-    ) {
-      const newConfig = {
-        ...this._config,
-        allergen_color_mode: value,
-        allergen_colors: LEVELS_DEFAULTS.allergen_colors,
-        allergen_outline_color: LEVELS_DEFAULTS.levels_gap_color,
-        no_allergens_color: LEVELS_DEFAULTS.no_allergens_color,
-      };
-      this._config = newConfig;
-      this._userConfig.allergen_color_mode = value;
-      delete this._userConfig.allergen_colors;
-      delete this._userConfig.allergen_outline_color;
-      delete this._userConfig.no_allergens_color;
-
-      this.dispatchEvent(
-        new CustomEvent("config-changed", {
-          detail: { config: newConfig },
-          bubbles: true,
-          composed: true,
-        }),
-      );
-      return;
     }
 
     // Specialfall: språkbyte – tvinga sort-dropdown att ritas om
@@ -1845,10 +1659,6 @@ class PollenPrognosCardEditor extends PollenEditorBase {
     const c = this._editorConfig();
     const allergens = this._currentAllergens();
     const numLevels = this._currentNumLevels();
-
-    const inheritMode = c.levels_inherit_mode || "inherit_allergen";
-    const gapSynced = c.allergen_levels_gap_synced ?? true;
-    const gapDisabled = inheritMode === "inherit_allergen" && gapSynced;
 
     if (this.debug) {
       console.debug("[Editor] Current language (lang):", this._lang);
@@ -2042,773 +1852,11 @@ class PollenPrognosCardEditor extends PollenEditorBase {
         ${this._renderAppearanceSection()}
 
         <!-- §6 Allergen icons -->
-        <details>
-          <summary>${this._t("summary_allergen_icons")}</summary>
-          <div class="section-helper">${this._t("helper_allergen_icons")}</div>
-          <ha-formfield
-            label="${this._t("allergen_color_mode") ||
-            "Allergen Color Mode"}"
-          >
-            <ha-selector
-              .hass=${this._hass}
-              .selector=${{
-                select: {
-                  mode: "dropdown",
-                  options: [
-                    {
-                      value: "default_colors",
-                      label:
-                        this._t("allergen_color_default_colors") ||
-                        "Default Colors",
-                    },
-                    {
-                      value: "custom",
-                      label:
-                        this._t("allergen_color_custom") || "Custom Colors",
-                    },
-                  ],
-                },
-              }}
-              .value=${c.allergen_color_mode || "default_colors"}
-              @value-changed=${(e) => {
-                const v = e.detail?.value;
-                if (v !== undefined)
-                  this._updateConfig("allergen_color_mode", v);
-              }}
-            ></ha-selector>
-          </ha-formfield>
-
-              ${c.allergen_color_mode === "custom"
-                ? html`
-                    <ha-formfield
-                      label="${this._t("allergen_colors") ||
-                      "Allergen Colors (by Level)"}"
-                    >
-                      <div
-                        style="display: flex; flex-direction: column; gap: 8px;"
-                      >
-                        ${(() => {
-                          // Use centralized default allergen colors from LEVELS_DEFAULTS
-                          const defaultAllergenColors =
-                            LEVELS_DEFAULTS.allergen_colors;
-                          const allergenColors =
-                            c.allergen_colors || defaultAllergenColors;
-
-                          return allergenColors.map(
-                            (col, i) => html`
-                              <div
-                                style="display: flex; align-items: center; gap: 8px;"
-                              >
-                                <span style="min-width: 60px;"
-                                  >Level ${i}:</span
-                                >
-                                <input
-                                  type="color"
-                                  .value=${(() => {
-                                    // For level 0, use a gray color for the preview since HTML color input can't show rgba
-                                    if (i === 0 && col.includes("rgba")) {
-                                      return "#c8c8c8"; // Gray equivalent of rgba(200,200,200,0.15)
-                                    }
-                                    return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                                      col,
-                                    )
-                                      ? col
-                                      : "#000000";
-                                  })()}
-                                  @input=${(e) => {
-                                    const newColors = [...allergenColors];
-                                    newColors[i] = e.target.value;
-                                    this._updateConfig(
-                                      "allergen_colors",
-                                      newColors,
-                                    );
-                                  }}
-                                  style="width: 28px; height: 28px; border: none; background: none;"
-                                />
-                                <ha-textfield
-                                  .value=${col}
-                                  placeholder="${i === 0
-                                    ? this._t("allergen_empty_placeholder") ||
-                                      "rgba(200,200,200,0.15)"
-                                    : this._t("allergen_colors_placeholder") ||
-                                      "#ffcc00"}"
-                                  @input=${(e) => {
-                                    const newColors = [...allergenColors];
-                                    newColors[i] = e.target.value;
-                                    this._updateConfig(
-                                      "allergen_colors",
-                                      newColors,
-                                    );
-                                  }}
-                                  style="width: 120px;"
-                                ></ha-textfield>
-                                <ha-button
-                                  outlined
-                                  title="${this._t("allergen_colors_reset") ||
-                                  "Reset"}"
-                                  @click=${() => {
-                                    const newColors = [...allergenColors];
-                                    newColors[i] =
-                                      LEVELS_DEFAULTS.allergen_colors[i];
-                                    this._updateConfig(
-                                      "allergen_colors",
-                                      newColors,
-                                    );
-                                  }}
-                                  style="margin-left: 8px;"
-                                  >↺</ha-button
-                                >
-                              </div>
-                            `,
-                          );
-                        })()}
-                      </div>
-                    </ha-formfield>
-
-                    <ha-formfield
-                      label="${this._t("no_allergens_color") ||
-                      "No Allergens Color"}"
-                    >
-                      <div
-                        style="display: flex; align-items: center; gap: 8px;"
-                      >
-                        <input
-                          type="color"
-                          .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                            c.no_allergens_color ||
-                              LEVELS_DEFAULTS.no_allergens_color,
-                          )
-                            ? c.no_allergens_color ||
-                              LEVELS_DEFAULTS.no_allergens_color
-                            : "#a9cfe0"}
-                          @input=${(e) =>
-                            this._updateConfig(
-                              "no_allergens_color",
-                              e.target.value,
-                            )}
-                          style="width: 28px; height: 28px; border: none; background: none;"
-                        />
-                        <ha-textfield
-                          .value=${c.no_allergens_color ||
-                          LEVELS_DEFAULTS.no_allergens_color}
-                          placeholder="${this._t(
-                            "no_allergens_color_placeholder",
-                          ) || "#a9cfe0"}"
-                          @input=${(e) =>
-                            this._updateConfig(
-                              "no_allergens_color",
-                              e.target.value,
-                            )}
-                          style="width: 100px;"
-                        ></ha-textfield>
-                        <ha-button
-                          outlined
-                          title="${this._t("no_allergens_color_reset") ||
-                          "Reset"}"
-                          @click=${() =>
-                            this._updateConfig(
-                              "no_allergens_color",
-                              LEVELS_DEFAULTS.no_allergens_color,
-                            )}
-                          style="margin-left: 8px;"
-                          >↺</ha-button
-                        >
-                      </div>
-                    </ha-formfield>
-                  `
-                : ""}
-          <ha-formfield
-            label="${this._t("allergen_outline_color")}"
-          >
-            <div style="display: flex; align-items: center; gap: 8px;">
-              <input
-                type="color"
-                .value=${(() => {
-                  const color =
-                    c.allergen_outline_color ||
-                    LEVELS_DEFAULTS.levels_gap_color;
-                  if (color.includes("rgba")) {
-                    return "#c8c8c8";
-                  }
-                  return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
-                    ? color
-                    : "#c8c8c8";
-                })()}
-                @input=${(e) =>
-                  this._updateConfig(
-                    "allergen_outline_color",
-                    e.target.value,
-                  )}
-                style="width: 28px; height: 28px; border: none; background: none;"
-              />
-              <ha-textfield
-                .value=${c.allergen_outline_color ||
-                LEVELS_DEFAULTS.levels_gap_color}
-                placeholder="${this._t(
-                  "allergen_outline_placeholder",
-                ) || "rgba(200,200,200,1)"}"
-                @input=${(e) =>
-                  this._updateConfig(
-                    "allergen_outline_color",
-                    e.target.value,
-                  )}
-                style="width: 100px;"
-              ></ha-textfield>
-              <ha-button
-                outlined
-                title="${this._t("allergen_outline_reset") || "Reset"}"
-                @click=${() =>
-                  this._updateConfig(
-                    "allergen_outline_color",
-                    LEVELS_DEFAULTS.levels_gap_color,
-                  )}
-                style="margin-left: 8px;"
-                >↺</ha-button
-              >
-            </div>
-          </ha-formfield>
-          <ha-formfield
-            label="${this._t("allergen_stroke_color_synced")}"
-          >
-            <ha-checkbox
-              .checked=${c.allergen_stroke_color_synced ?? true}
-              @change=${(e) =>
-                this._updateConfig(
-                  "allergen_stroke_color_synced",
-                  e.target.checked,
-                )}
-            ></ha-checkbox>
-          </ha-formfield>
-          <ha-formfield
-            label="${this._t("allergen_stroke_width")}"
-          >
-            <ha-slider
-              min="0"
-              max="150"
-              step="5"
-              .value=${c.allergen_stroke_width ?? LEVELS_DEFAULTS.allergen_stroke_width}
-              @input=${(e) => {
-                const value = Number(e.target.value);
-                this._updateConfig("allergen_stroke_width", value);
-                if (
-                  (c.levels_inherit_mode || "inherit_allergen") ===
-                    "inherit_allergen" &&
-                  (c.allergen_levels_gap_synced ?? true)
-                ) {
-                  const levelGap = convertStrokeWidthToGap(value);
-                  this._updateConfig("levels_gap", levelGap);
-                }
-              }}
-              style="width: 120px;"
-            ></ha-slider>
-            <ha-textfield
-              type="number"
-              min="0"
-              max="150"
-              step="5"
-              .value=${c.allergen_stroke_width ?? LEVELS_DEFAULTS.allergen_stroke_width}
-              @input=${(e) => {
-                const value = e.target.value === "" ? LEVELS_DEFAULTS.allergen_stroke_width : Number(e.target.value);
-                this._updateConfig("allergen_stroke_width", value);
-                if (
-                  (c.levels_inherit_mode || "inherit_allergen") ===
-                    "inherit_allergen" &&
-                  (c.allergen_levels_gap_synced ?? true)
-                ) {
-                  const levelGap = convertStrokeWidthToGap(value);
-                  this._updateConfig("levels_gap", levelGap);
-                }
-              }}
-              style="width: 80px;"
-            ></ha-textfield>
-            <ha-button
-              outlined
-              title="${this._t("allergen_stroke_width_reset") || "Reset"}"
-              @click=${() =>
-                this._updateConfig(
-                  "allergen_stroke_width",
-                  LEVELS_DEFAULTS.allergen_stroke_width,
-                )}
-              style="margin-left: 8px;"
-              >↺</ha-button
-            >
-          </ha-formfield>
-          <div class="field-helper">${this._t("helper_allergen_stroke_width")}</div>
-        </details>
-
+        ${this._renderAllergenIconsSection()}
         <!-- §7 Level circles -->
-        <details>
-          <summary>${this._t("summary_level_circles")}</summary>
-          <div class="section-helper">${this._t("helper_level_circles")}</div>
-          <ha-formfield
-            label="${this._t("levels_inherit_mode")}"
-          >
-            <ha-selector
-              .hass=${this._hass}
-              .selector=${{
-                select: {
-                  mode: "dropdown",
-                  options: [
-                    {
-                      value: "inherit_allergen",
-                      label:
-                        this._t("levels_inherit_allergen") ||
-                        "Inherit from Allergen Colors",
-                    },
-                    {
-                      value: "custom",
-                      label:
-                        this._t("levels_custom") || "Use Custom Level Colors",
-                    },
-                  ],
-                },
-              }}
-              .value=${c.levels_inherit_mode || "inherit_allergen"}
-              @value-changed=${(e) => {
-                const v = e.detail?.value;
-                if (v !== undefined) this._updateConfig("levels_inherit_mode", v);
-              }}
-            ></ha-selector>
-          </ha-formfield>
-
-          ${inheritMode === "inherit_allergen"
-            ? html`
-                <ha-formfield
-                  label="${this._t("allergen_levels_gap_synced")}"
-                >
-                  <ha-checkbox
-                    .checked=${c.allergen_levels_gap_synced ?? true}
-                    @change=${(e) =>
-                      this._updateConfig(
-                        "allergen_levels_gap_synced",
-                        e.target.checked,
-                      )}
-                  ></ha-checkbox>
-                </ha-formfield>
-                <div class="field-helper">${this._t("helper_allergen_levels_gap_synced")}</div>
-              `
-            : ""}
-
-          ${inheritMode === "custom"
-            ? html`
-                <ha-formfield label="${this._t("levels_colors")}">
-                  <div style="display: flex; flex-direction: column; gap: 8px;">
-                    ${c.levels_colors.map(
-                      (col, i) => html`
-                        <div style="display: flex; align-items: center; gap: 8px;">
-                          <input
-                            type="color"
-                            .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(col)
-                              ? col
-                              : "#000000"}
-                            @input=${(e) => {
-                              const newColors = [...c.levels_colors];
-                              newColors[i] = e.target.value;
-                              this._updateConfig("levels_colors", newColors);
-                            }}
-                            style="width: 28px; height: 28px; border: none; background: none;"
-                          />
-                          <ha-textfield
-                            .value=${col}
-                            placeholder="${this._t("levels_colors_placeholder")}"
-                            @input=${(e) => {
-                              const newColors = [...c.levels_colors];
-                              newColors[i] = e.target.value;
-                              this._updateConfig("levels_colors", newColors);
-                            }}
-                            style="width: 100px;"
-                          ></ha-textfield>
-                          <ha-button
-                            outlined
-                            title="${this._t("levels_reset")}"
-                            @click=${() => {
-                              const newColors = [...c.levels_colors];
-                              newColors[i] = LEVELS_DEFAULTS.levels_colors[i];
-                              this._updateConfig("levels_colors", newColors);
-                            }}
-                            style="margin-left: 8px;"
-                            >↺</ha-button
-                          >
-                        </div>
-                      `,
-                    )}
-                  </div>
-                </ha-formfield>
-
-                <ha-formfield label="${this._t("levels_empty_color")}">
-                  <div style="display: flex; align-items: center; gap: 8px;">
-                    <input
-                      type="color"
-                      .value=${(() => {
-                        const color =
-                          c.levels_empty_color ||
-                          LEVELS_DEFAULTS.levels_empty_color;
-                        if (color.includes("rgba")) {
-                          return "#c8c8c8";
-                        }
-                        return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
-                          ? color
-                          : "#c8c8c8";
-                      })()}
-                      @input=${(e) =>
-                        this._updateConfig("levels_empty_color", e.target.value)}
-                      style="width: 28px; height: 28px; border: none; background: none;"
-                    />
-                    <ha-textfield
-                      .value=${c.levels_empty_color}
-                      placeholder="${this._t("levels_colors_placeholder")}"
-                      @input=${(e) =>
-                        this._updateConfig("levels_empty_color", e.target.value)}
-                      style="width: 100px;"
-                    ></ha-textfield>
-                    <ha-button
-                      outlined
-                      title="${this._t("levels_reset")}"
-                      @click=${() =>
-                        this._updateConfig(
-                          "levels_empty_color",
-                          LEVELS_DEFAULTS.levels_empty_color,
-                        )}
-                      style="margin-left: 8px;"
-                      >↺</ha-button
-                    >
-                  </div>
-                </ha-formfield>
-              `
-            : ""}
-
-          <ha-formfield label="${this._t("levels_thickness")}">
-            <ha-slider
-              min="10"
-              max="90"
-              step="1"
-              .value=${c.levels_thickness}
-              @input=${(e) =>
-                this._updateConfig("levels_thickness", Number(e.target.value))}
-              style="width: 120px;"
-            ></ha-slider>
-            <ha-textfield
-              type="number"
-              .value=${c.levels_thickness}
-              @input=${(e) =>
-                this._updateConfig("levels_thickness", Number(e.target.value))}
-              style="width: 80px;"
-            ></ha-textfield>
-            <ha-button
-              outlined
-              title="${this._t("levels_reset")}"
-              @click=${() =>
-                this._updateConfig("levels_thickness", LEVELS_DEFAULTS.levels_thickness)}
-              style="margin-left: 8px;"
-              >↺</ha-button
-            >
-          </ha-formfield>
-
-          <ha-formfield
-            label="${this._t("levels_gap")}"
-            .disabled=${gapDisabled}
-          >
-            <ha-slider
-              min="0"
-              max="20"
-              step="1"
-              .value=${c.levels_gap}
-              .disabled=${gapDisabled}
-              @input=${(e) =>
-                this._updateConfig("levels_gap", Number(e.target.value))}
-              style="width: 120px;"
-            ></ha-slider>
-            <ha-textfield
-              type="number"
-              .value=${c.levels_gap}
-              .disabled=${gapDisabled}
-              @input=${(e) =>
-                this._updateConfig("levels_gap", Number(e.target.value))}
-              style="width: 80px;"
-            ></ha-textfield>
-            <ha-button
-              outlined
-              title="${this._t("levels_reset")}"
-              .disabled=${gapDisabled}
-              @click=${() =>
-                this._updateConfig("levels_gap", LEVELS_DEFAULTS.levels_gap)}
-              style="margin-left: 8px;"
-              >↺</ha-button
-            >
-          </ha-formfield>
-          <div class="field-helper">
-            ${gapDisabled
-              ? this._t("helper_levels_gap_synced")
-              : this._t("helper_levels_gap_unsynced")}
-          </div>
-
-          ${inheritMode === "custom" || !gapSynced
-            ? html`
-                <ha-formfield label="${this._t("levels_gap_color")}">
-                  <div style="display: flex; align-items: center; gap: 8px;">
-                    <input
-                      type="color"
-                      .value=${(() => {
-                        const color =
-                          c.levels_gap_color || LEVELS_DEFAULTS.levels_gap_color;
-                        if (color.includes("rgba")) {
-                          return "#c8c8c8";
-                        }
-                        return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
-                          ? color
-                          : "#c8c8c8";
-                      })()}
-                      @input=${(e) =>
-                        this._updateConfig("levels_gap_color", e.target.value)}
-                      style="width: 28px; height: 28px; border: none; background: none;"
-                    />
-                    <ha-textfield
-                      .value=${c.levels_gap_color}
-                      placeholder="${this._t("levels_colors_placeholder")}"
-                      @input=${(e) =>
-                        this._updateConfig("levels_gap_color", e.target.value)}
-                      style="width: 100px;"
-                    ></ha-textfield>
-                    <ha-button
-                      outlined
-                      title="${this._t("levels_reset")}"
-                      @click=${() =>
-                        this._updateConfig(
-                          "levels_gap_color",
-                          LEVELS_DEFAULTS.levels_gap_color,
-                        )}
-                      style="margin-left: 8px;"
-                      >↺</ha-button
-                    >
-                  </div>
-                </ha-formfield>
-              `
-            : ""}
-
-          <ha-formfield label="${this._t("show_value_numeric_in_circle")}">
-            <ha-switch
-              .checked=${c.show_value_numeric_in_circle}
-              @change=${(e) =>
-                this._updateConfig(
-                  "show_value_numeric_in_circle",
-                  e.target.checked,
-                )}
-            ></ha-switch>
-          </ha-formfield>
-          <div class="field-helper">${this._t("helper_show_value_numeric_in_circle")}</div>
-
-          <ha-formfield label="${this._t("levels_text_weight")}">
-            <ha-selector
-              .hass=${this._hass}
-              .selector=${{
-                select: {
-                  mode: "dropdown",
-                  options: [
-                    { value: "normal", label: "Normal" },
-                    { value: "500", label: "Medium" },
-                    { value: "bold", label: "Bold" },
-                  ],
-                },
-              }}
-              .value=${c.levels_text_weight || "normal"}
-              @value-changed=${(e) => {
-                const v = e.detail?.value;
-                if (v !== undefined) this._updateConfig("levels_text_weight", v);
-              }}
-            ></ha-selector>
-          </ha-formfield>
-
-          <ha-formfield label="${this._t("levels_text_size")}">
-            <ha-slider
-              min="0.1"
-              max="0.5"
-              step="0.05"
-              .value=${c.levels_text_size || 0.3}
-              @input=${(e) =>
-                this._updateConfig("levels_text_size", Number(e.target.value))}
-              style="width: 120px;"
-            ></ha-slider>
-            <ha-textfield
-              type="number"
-              .value=${c.levels_text_size || 0.3}
-              @input=${(e) =>
-                this._updateConfig("levels_text_size", Number(e.target.value))}
-              style="width: 80px;"
-            ></ha-textfield>
-          </ha-formfield>
-
-          <ha-formfield label="${this._t("levels_icon_ratio")}">
-            <ha-slider
-              min="0.1"
-              max="2"
-              step="0.05"
-              .value=${c.levels_icon_ratio || 1}
-              @input=${(e) =>
-                this._updateConfig("levels_icon_ratio", Number(e.target.value))}
-              style="width: 120px;"
-            ></ha-slider>
-            <ha-textfield
-              type="number"
-              .value=${c.levels_icon_ratio || 1}
-              @input=${(e) =>
-                this._updateConfig("levels_icon_ratio", Number(e.target.value))}
-              style="width: 80px;"
-            ></ha-textfield>
-          </ha-formfield>
-
-          <ha-formfield label="${this._t("levels_text_color")}">
-            <div style="display: flex; align-items: center; gap: 8px;">
-              <input
-                type="color"
-                .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                  c.levels_text_color || "",
-                )
-                  ? c.levels_text_color
-                  : "#000000"}
-                @input=${(e) =>
-                  this._updateConfig("levels_text_color", e.target.value)}
-                style="width: 28px; height: 28px; border: none; background: none;"
-              />
-              <ha-textfield
-                .value=${c.levels_text_color || ""}
-                placeholder="var(--primary-text-color)"
-                @input=${(e) =>
-                  this._updateConfig("levels_text_color", e.target.value)}
-                style="width: 100px;"
-              ></ha-textfield>
-            </div>
-          </ha-formfield>
-        </details>
-
+        ${this._renderLevelCirclesSection()}
         <!-- §8 Icon in ring -->
-        <details>
-          <summary>${this._t("summary_icon_in_ring")}</summary>
-          <div class="section-helper">${this._t("helper_icon_in_ring")}</div>
-          <ha-formfield
-            label="${this._t("icon_in_ring")}"
-          >
-            <ha-checkbox
-              .checked=${c.icon_in_ring === true}
-              @change=${(e) =>
-                this._updateConfig("icon_in_ring", e.target.checked)}
-            ></ha-checkbox>
-          </ha-formfield>
-          <ha-formfield
-            label="${this._t("icon_in_ring_size_ratio")}"
-          >
-            <div style="display: flex; align-items: center; gap: 8px;">
-              <ha-slider
-                min="0.2"
-                max="0.9"
-                step="0.05"
-                .value=${c.icon_in_ring_size_ratio ??
-                LEVELS_DEFAULTS.icon_in_ring_size_ratio}
-                @change=${(e) =>
-                  this._updateConfig(
-                    "icon_in_ring_size_ratio",
-                    Number(e.target.value),
-                  )}
-              ></ha-slider>
-              <ha-textfield
-                type="number"
-                min="0.2"
-                max="0.9"
-                step="0.05"
-                .value=${c.icon_in_ring_size_ratio ??
-                LEVELS_DEFAULTS.icon_in_ring_size_ratio}
-                @change=${(e) =>
-                  this._updateConfig(
-                    "icon_in_ring_size_ratio",
-                    Number(e.target.value),
-                  )}
-                style="width: 80px;"
-              ></ha-textfield>
-            </div>
-          </ha-formfield>
-          <ha-formfield
-            label="${this._t("icon_in_ring_color_mode")}"
-          >
-            <ha-selector
-              .hass=${this._hass}
-              .selector=${{
-                select: {
-                  mode: "dropdown",
-                  options: [
-                    {
-                      value: "static",
-                      label:
-                        this._t("icon_in_ring_color_static") ||
-                        "Static color",
-                    },
-                    {
-                      value: "follow_level",
-                      label:
-                        this._t("icon_in_ring_color_follow") ||
-                        "Follow level color",
-                    },
-                  ],
-                },
-              }}
-              .value=${c.icon_in_ring_color_mode || "static"}
-              @value-changed=${(e) => {
-                const v = e.detail?.value;
-                if (v !== undefined)
-                  this._updateConfig("icon_in_ring_color_mode", v);
-              }}
-            ></ha-selector>
-          </ha-formfield>
-          ${(c.icon_in_ring_color_mode || "static") === "static"
-            ? (() => {
-                const currentColor =
-                  typeof c.icon_in_ring_static_color === "string"
-                    ? c.icon_in_ring_static_color
-                    : LEVELS_DEFAULTS.icon_in_ring_static_color;
-                const isHex = /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                  currentColor,
-                );
-                return html`
-                  <ha-formfield
-                    label="${this._t("icon_in_ring_static_color")}"
-                  >
-                    <div
-                      style="display: flex; align-items: center; gap: 8px;"
-                    >
-                      ${isHex
-                        ? html`<input
-                            type="color"
-                            .value=${currentColor}
-                            @input=${(e) =>
-                              this._updateConfig(
-                                "icon_in_ring_static_color",
-                                e.target.value,
-                              )}
-                            style="width: 28px; height: 28px; border: none; background: none;"
-                          />`
-                        : html`<div
-                            title="${this._t(
-                              "icon_in_ring_static_color_var_hint",
-                            ) ||
-                            "Non-hex value (e.g. CSS variable); use the text field to edit"}"
-                            style="width: 28px; height: 28px; border: 1px dashed var(--divider-color); border-radius: 4px; display: flex; align-items: center; justify-content: center; font-size: 10px; color: var(--secondary-text-color);"
-                          >
-                            var
-                          </div>`}
-                      <ha-textfield
-                        .value=${currentColor}
-                        placeholder="${LEVELS_DEFAULTS.icon_in_ring_static_color}"
-                        @input=${(e) =>
-                          this._updateConfig(
-                            "icon_in_ring_static_color",
-                            e.target.value,
-                          )}
-                        style="width: 220px;"
-                      ></ha-textfield>
-                    </div>
-                  </ha-formfield>
-                `;
-              })()
-            : ""}
-        </details>
+        ${this._renderIconInRingSection()}
 
         <!-- §9 Translations & strings -->
         <details>
